+++ Sources/Rapid/Strings/Interpolation.swift
@@ -33,4 +33,27 @@ public extension String.StringInterpolation {
     mutating func appendInterpolation<Style: FormatStyle>(_ value: Style.FormatInput, format formatStyle: Style) {
         appendInterpolation(formatStyle.format(value))
     }
+    
+    /// Interpolates the given value into the string literal being created
+    /// if the provided condition is `true`.
+    ///
+    /// Do not call this method directly. It is used by the compiler when
+    /// interpreting string interpolations. Instead, use string interpolation
+    /// to create a new string by including values, literals, variables, or
+    /// expressions enclosed in parentheses, prefixed by a backslash
+    /// (`\(`...`)`).
+    ///
+    /// This method uses short-circuit evaluation: The condition is evaluated
+    /// first, and the value to interpolate is evaluated only if the condition
+    /// evaluates to `true`.
+    ///
+    /// - Parameters:
+    ///   - value: The value to interpolate if `condition` is `true`.
+    ///   - condition: A Boolean condition.
+    mutating func appendInterpolation(_ value: @autoclosure () -> StringLiteralType, if condition: Bool) {
+        guard condition else {
+            return
+        }
+        appendLiteral(value())
+    }
 }